I got an A+ on my summative Lang synthesis essay

+++ src/overseer.h
@@ -10,11 +10,11 @@ struct vec2
 
 int magnitude (const int tx, const int ty, const vec2 motor);
 
-int curve (int input);
+int curve (float input);
 
 struct Controls
 {
-	int forward, right, up, clockwise, moclaw;
+	float forward, right, up, clockwise, moclaw;
 };
 
 void serialize_controls(std::array<unsigned char, 12>& pin_data, const Controls& c, int botflag);
